refactor: refactoring paste-button
