+++ src/course-outline/CourseOutline.scss
@@ -8,4 +8,3 @@
 @import "./highlights-modal/HighlightsModal";
 @import "./publish-modal/PublishModal";
 @import "./xblock-status/XBlockStatus";
-@import "./paste-button/PasteButton";
