--- src/course-outline/paste-button/PasteButton.scss
@@ -1,20 +1,0 @@
-// adds bottom arrow to popup link
-.popup-link {
-  position: relative;
-
-  &::after {
-    content: "";
-    position: absolute;
-    top: 100%;
-    left: 50%;
-    width: 0;
-    height: 0;
-    border-top: solid .5rem white;
-    border-left: solid .5rem transparent;
-    border-right: solid .5rem transparent;
-  }
-}
-
-.cursor-help {
-  cursor: help !important;
-}
