--- src/course-outline/paste-button/messages.js
@@ -1,14 +1,0 @@
-import { defineMessages } from '@edx/frontend-platform/i18n';
-
-const messages = defineMessages({
-  clipboardContentFromLabel: {
-    id: 'course-authoring.course-outline.paste-button.whats-in-clipboard.from-label',
-    defaultMessage: 'From: ',
-  },
-  clipboardContentLabel: {
-    id: 'course-authoring.course-outline.paste-button.whats-in-clipboard.label',
-    defaultMessage: 'What\'s in my clipboard?',
-  },
-});
-
-export default messages;
