+++ src/course-unit/CourseUnit.scss
@@ -3,5 +3,4 @@
 @import "./add-component/AddComponent";
 @import "./sidebar/Sidebar";
 @import "./course-xblock/CourseXblock";
-@import "./clipboard/paste-component/PasteComponent";
 @import "./header-title/HeaderTitle";
