+++ src/course-unit/clipboard/index.js
@@ -1,3 +1,2 @@
-export { default as PasteComponent } from './paste-component';
 export { default as PasteNotificationAlert } from './paste-notification';
 export { default as useCopyToClipboard } from './hooks/useClipboard';
