+++ src/generic/paste-button/PasteButton.scss
