+++ src/generic/paste-button/components/PopoverContent.jsx
