+++ src/generic/paste-button/components/WhatsInClipboard.jsx
