+++ src/generic/paste-button/components/index.js
