+++ src/generic/styles.scss
@@ -10,3 +10,4 @@
 @import "./configure-modal/ConfigureModal";
 @import "./drag-helper/ConditionalSortableElement";
 @import "./divider/Divider";
+@import "./paste-button/PasteButton";
